Aj-87(Added real-time asynchronous caching progress updates from the service worker!)
I have Added  real-time asynchronous caching progress updates from the service worker just as ftmpwa,
Since there is not much difference in the way we are caching the content in ftmpwa and assessment, that's why I have only done few modification only and updated Service!

Along with that the thing to  note here is in my research we are using workbox to precache the content whose actual percantage we can't calculate and using the service worker for caching all the content will also be an hindrance in precaching all the content and sending automated Updates!